# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TMJConnect API
  version: 1.0.0
  description: |
    HIPAA-compliant orofacial pain management platform API.
    Supports patients, providers, and admin roles.
  contact:
    name: TMJConnect Engineering
servers:
  - url: /api/v1
    description: API v1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Notification:
      description: In-app notification record
      type: object
      required: [id, type, title, body, read, data, created_at]
      properties:
        id: { type: string, format: uuid }
        type:
          type: string
          description: "Notification type (determines channel routing)"
          enum: [exercise_reminder, symptom_checkin, provider_message, report_submitted, report_urgent, report_reviewed, exercise_assigned, link_accepted, welcome, password_reset, mfa_code, new_device_login, weekly_summary, account_locked, streak_milestone]
        title: { type: string }
        body: { type: string }
        read: { type: boolean }
        data:
          $ref: '#/components/schemas/NotificationData'
        created_at: { type: string, format: date-time }
    NotificationData:
      description: Extra payload for deep linking and context
      type: object
      properties:
        reportId: { type: string, format: uuid, description: "For report-related notifications" }
        reminderId: { type: string, format: uuid, description: "For reminder notifications" }
        patientId: { type: string, format: uuid, description: "For provider-facing notifications" }
        providerId: { type: string, format: uuid, description: "For patient-facing link notifications" }
        urgency: { type: string, enum: [routine, concerning, urgent], description: "Report urgency level" }
        action: { type: string, description: "Sub-action (e.g. 'responded')" }
      additionalProperties: true
...
    LinkingCode:
      description: Provider-generated 6-character patient invite code (7-day expiry)
      type: object
      required: [id, code, provider_id, status, expires_at, created_at]
      properties:
        id: { type: string, format: uuid }
        code: { type: string, minLength: 6, maxLength: 6, description: "6-char alphanumeric code" }
        provider_id: { type: string, format: uuid }
        patient_id: { type: [string, "null"], format: uuid, description: "Set when patient accepts" }
        status: { type: string, enum: [pending, connected, expired] }
        expires_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
    Link:
      description: Active patient-provider connection
      type: object
      required: [link_id, first_name, last_name, linked_at]
      properties:
        link_id: { type: string, format: uuid }
        patient_id: { type: string, format: uuid, description: "Present in provider's view" }
        provider_id: { type: string, format: uuid, description: "Present in patient's view" }
        first_name: { type: string, description: "Name of the other party" }
        last_name: { type: string }
        linked_at: { type: string, format: date-time }

    # ─── Uploads ───────────────────────────────────────────────────────────────
...
    PageMeta:
      description: Offset-based pagination metadata
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page: { type: integer, description: "Current page number" }
        limit: { type: integer, description: "Items per page" }
        total: { type: integer, description: "Total matching items" }
        totalPages: { type: integer, description: "Total pages" }
    CursorMeta:
      description: Cursor-based pagination metadata
      type: object
      required: [hasMore]
      properties:
        nextCursor: { type: [string, "null"], description: "Pass as cursor param for next page" }
        hasMore: { type: boolean }

    # ─── Provider Dashboard ────────────────────────────────────────────────────
...
    LinkingCodeRow:
      description: Linking code with the issuing provider's name (admin view)
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        provider_id: { type: string, format: uuid }
        provider_name: { type: string }
        status: { type: string, enum: [pending, accepted, expired, revoked] }
        invited_email: { type: [string, "null"], format: email }
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        accepted_at: { type: [string, "null"], format: date-time }
    LinkingLinkRow:
      description: Patient–provider link with both names (admin view)
      type: object
      properties:
        id: { type: string, format: uuid }
        patient_id: { type: string, format: uuid }
        patient_name: { type: string }
        provider_id: { type: string, format: uuid }
        provider_name: { type: string }
        linked_at: { type: string, format: date-time }
        unlinked_at: { type: [string, "null"], format: date-time }
...
paths:
...
  /notifications:
    get:
      tags: [Notifications]
      summary: List notifications (cursor paginated)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: cursor
          in: query
          schema: { type: string, format: date-time }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Notification' }
                  meta: { $ref: '#/components/schemas/CursorMeta' }
                  unread_count: { type: integer }
  /notifications/read-all:
    patch:
      tags: [Notifications]
      summary: Mark all notifications as read
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: All notifications marked read }
  /notifications/{id}/read:
    patch:
      tags: [Notifications]
      summary: Mark a single notification as read
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Notification' }

  # ──────────────────────────── REMINDERS ────────────────────────────
...
  /linking/links:
    get:
      tags: [Linking]
      summary: List active links (both roles)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Link' }
  /linking/links/{linkId}:
    delete:
      tags: [Linking]
      summary: Disconnect a link (either party)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: linkId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Link disconnected }
  /linking/codes:
    post:
      tags: [Linking]
      summary: Generate a linking code (provider)
      security: [{ bearerAuth: [] }]
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/LinkingCode' }
    get:
      tags: [Linking]
      summary: List generated codes (provider)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/LinkingCode' }
  /linking/codes/{codeId}/invite:
    post:
      tags: [Linking]
      summary: Send email invite with linking code (provider)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: codeId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [patient_email]
              properties:
                patient_email: { type: string, format: email }
                patient_name: { type: string, maxLength: 200 }
      responses:
        '202': { description: Invitation email queued }
  /linking/accept:
    post:
      tags: [Linking]
      summary: Accept a linking code (patient)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string, minLength: 6, maxLength: 6 }
      responses:
        '201': { description: Link created. Both parties notified. }
        '404': { description: Invalid or expired code }
        '409': { description: Already linked to this provider }
        '410': { description: Code expired }

  # ──────────────────────────── UPLOADS ────────────────────────────
...
  /admin/linking/codes:
    get:
      tags: [Admin]
      summary: List linking codes with provider names
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { name: status, in: query, schema: { type: string, enum: [pending, accepted, expired, revoked] } }
      responses:
        '200':
          description: Codes
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/LinkingCodeRow' }
                  meta: { $ref: '#/components/schemas/PageMeta' }
  /admin/linking/links:
    get:
      tags: [Admin]
      summary: List patient–provider links (active or disconnected)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { name: status, in: query, schema: { type: string, enum: [active, disconnected] } }
      responses:
        '200':
          description: Links
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/LinkingLinkRow' }
                  meta: { $ref: '#/components/schemas/PageMeta' }
...